ioremap: Converting all device drivers to use mmio pointer
Device drivers need to use pointer based
mmio reads/writes instead of using static value. This protects
their working after switching to high memory execution.

+++ gpio.c
@@ -40,7 +40,7 @@ void init_gpio(struct gpio_controller *ctrl)
 			return;
 	}
 
-	ctrl->base = (uint64_t *) ioremap(mmio_base + base);
+	ctrl->base = (uint64_t *) ioremap((uint64_t)mmio_base + base);
 
 	/* generic functions */
 	ctrl->ctrl_set_gpio_fn = generic_set_gpio_func;
